+++ release-from-jail/requirements.txt
@@ -1,2 +1,3 @@
-web3>=5.19.0
+web3==5.19.0
 requests>=2.31.0
+eth_abi==2.2.0
